SmallCollections.jl

SmallCollections — Module

SmallCollections

This packages provides several immutable collections that don't allocate and are therefore faster than the usual types. The number of elements that these collections can hold is necessarily limited. At present SmallBitSet and SmallVector are defined.

If the package BangBang.jl is loaded, then many functions defined by this package are also available in !!-form. For example, setindex!! with a SmallVector as first argument calls setindex.

See SmallBitSet, SmallVector.

SmallVector

SmallCollections.SmallVector — Type

SmallVector{N,T} <: AbstractVector{T}

SmallVector{N,T}(iter)
SmallVector{N}(v::AbstractVector{T})
SmallVector{N}(t::Tuple)

SmallVector{N,T} is an immutable vector type that can hold up to N elements of type T. Here N can be any (small) positive integer. However, at least for bit integer and hardware float types, one usually takes N to be a power of 2.

The element type T can be omitted when creating the SmallVector from an AbstractVector or from a tuple. In the latter case, T is determined by promoting the element types of the tuple.

The unused elements of a SmallVector{N,T} are filled with the value default(T). This is pre-defined for number types, Char, String and Symbol. For other types it must be set explicitly.

Addition and subtraction of two SmallVectors is possible even if the vectors have different capacity. (Of course, their lengths must agree.) The capacity of the result is the smaller of the arguments' capacities in this case.

See also capacity, SmallCollections.default, promote_type.

Examples

julia> v = SmallVector{8,Int8}(2*x for x in 1:3)
3-element SmallVector{8, Int8}:
 2
 4
 6

julia> w = SmallVector{9}((1, 2.5, 4))
3-element SmallVector{9, Float64}:
 1.0
 2.5
 4.0

julia> v+w
3-element SmallVector{8, Float64}:
  3.0
  6.5
 10.0

SmallCollections.capacity — Method

capacity(::Type{<:SmallVector{N}}) where N -> N
capacity(v::SmallVector{N}) where N -> N

Return N, which is the largest number of elements this vector type can hold.

SmallCollections.fasthash — Method

fasthash(v::SmallVector [, h0::UInt]) -> UInt

Return a hash for v that may be computed faster than the standard hash for vectors. This new hash is consistent across all SmallVectorss of the same element type, but it may not be compatible with hash or with fasthash for a SmallVector having a different element type.

Currently, fasthash differs from hash only if the element type of v is a bit integer type with at most 32 bits, Bool or Char.

See also Base.hash.

Examples

julia> v = SmallVector{8,Int8}([1, 5, 6]);

julia> fasthash(v)
0x6466067ab41d0916

julia> fasthash(v) == hash(v)
false

julia> w = SmallVector{16,Int8}(v); fasthash(v) == fasthash(w)
true

julia> w = SmallVector{8,Int16}(v); fasthash(v) == fasthash(w)
false

Base.empty — Function

empty(v::V) where V <: SmallVector -> V

Return an empty SmallVector of the same type as v.

Base.zeros — Method

zeros(::Type{V}, n::Integer) where V <: SmallVector -> V

Return a SmallVector of type V containing n zeros.

See also ones(::Type{<:SmallVector}, ::Integer).

Base.ones — Method

ones(::Type{V}, n::Integer) where V <: SmallVector -> V

Return a SmallVector of type V containing n ones.

See also zeros(::Type{<:SmallVector}, ::Integer).

Base.setindex — Function

setindex(v::V, x, i::Integer) where V <: SmallVector -> V

Return a vector that agrees with v except possibly for the i-th entry that is set to x.

See also Base.setindex.

SmallCollections.push — Method

push(v::SmallVector{N,T}, xs...) where {N,T} -> SmallVector{N,T}

Return the SmallVector obtained from v by appending the other arguments xs. The length of v must be less than N.

See also Base.push!, BangBang.push!!.

SmallCollections.pop — Method

pop(v::SmallVector{N,T}) where {N,T} -> Tuple{SmallVector{N,T},T}

Return the tuple (w, x) where x is the last element of v and w obtained from v by dropping this element. The vector v must not be empty.

See also Base.pop!, BangBang.pop!!.

SmallCollections.pushfirst — Function

pushfirst(v::SmallVector{N,T}, xs...) where {N,T} -> SmallVector{N,T}

Return the SmallVector obtained from v by prepending the other arguments xs. The length of v must be less than N.

See also Base.pushfirst!, BangBang.pushfirst!!.

SmallCollections.popfirst — Function

popfirst(v::SmallVector{N,T}) where {N,T} -> Tuple{SmallVector{N,T},T}

Return the tuple (w, x) where x is the first element of v and w obtained from v by dropping this element. The vector v must not be empty.

See also Base.popfirst!, BangBang.popfirst!!.

SmallCollections.insert — Function

insert(v::SmallVector{N,T}, i::Integer, x) where {N,T} -> SmallVector{N,T}

Return the SmallVector obtained from v by inserting x at position i. The position i must be between 1 and length(v)+1, and length(v) must be less than N.

This is the non-mutating analog of Base.insert!.

SmallCollections.deleteat — Function

deleteat(v::V, i::Integer) where V <: SmallVector -> V

Return the SmallVector obtained from v by deleting the element at position i. The position i must be between 1 and length(v).

See also Base.deleteat!, BangBang.deleteat!!.

SmallCollections.popat — Function

popat(v::SmallVector{N,T}, i::Integer) where {N,T} -> Tuple{SmallVector{N,T},T}

Return the tuple (w, x) where w obtained from v by deleting the element x at position i. The latter must be between 1 and length(v).

See also Base.popat!, BangBang.popat!!.

SmallCollections.support — Function

support(v::SmallVector) -> SmallBitSet

Return the SmallBitSet with the indices of the non-zero elements of v.

See also SmallBitSet.

Example

julia> v = SmallVector{8,Int8}([1, 0, 2, 0, 0, 3]);

julia> support(v)
SmallBitSet{UInt64} with 3 elements:
  1
  3
  6

SmallBitSet

SmallCollections.SmallBitSet — Type

SmallBitSet{U<:Unsigned} <: AbstractSet{Int}

SmallBitSet{U}([iter])
SmallBitSet([iter])

SmallBitSet{U} is an immutable set that can hold integers between 1 and the bit length of U. Called without an argument, it returns an empty set. If U is omitted, then UInt is taken.

All non-mutating functions for sets are supported. The non-mutating analogs push, pop and delete of the corresponding !-functions are also provided.

SmallCollections.bits — Function

bits(s::SmallBitSet{U}) where U -> U

Return the bit mask used internally to store the elements of the set s.

See also convert(::Type{SmallBitSet}, ::Integer).

Base.convert — Method

convert(::Type{SmallBitSet{U}}, mask::Integer) where U -> SmallBitSet{U}
convert(::Type{SmallBitSet}, mask::Integer) -> SmallBitSet{UInt}

Convert a bit mask to a SmallBitSet of the given type. This is the inverse operation to bits.

See also bits.

Examples

julia> s = SmallBitSet{UInt16}([1, 5, 6]);

julia> u = bits(s)
0x0031

julia> convert(SmallBitSet, u)
SmallBitSet{UInt64} with 3 elements:
  1
  5
  6

SmallCollections.capacity — Method

capacity(::Type{<:SmallBitSet}) -> Int
capacity(s::SmallBitSet) -> Int

Return the largest number that the given set or SmallBitSet type can store.

SmallCollections.fasthash — Method

fasthash(s::SmallBitSet [, h0::UInt]) -> UInt

Return a hash for s that can be computed fast. This hash is consistent across all SmallBitSets, but it is not compatible with the hash used for sets.

See also Base.hash.

Examples

julia> s = SmallBitSet(1:3);

julia> fasthash(s)
0x828a4cc485149963

julia> fasthash(s) == hash(s)
false

julia> t = SmallBitSet{UInt16}(s);

julia> fasthash(s) == fasthash(t)
true

SmallCollections.push — Method

push(s::S, xs...) where S <: SmallBitSet -> S

Return the SmallBitSet obtained from s by adding the other arguments xs.

See also Base.push!, BangBang.push!!.

SmallCollections.pop — Method

pop(s::S) where S <: SmallBitSet -> Tuple{S, Int}

Return the pair (t, x) where x is the smallest element from s and t is the set s with x deleted. The set s must be non-empty.

See also Base.pop!, BangBang.pop!!.

SmallCollections.pop — Method

pop(s::S, x) where S <: SmallBitSet -> Tuple{S, Int}

Return the pair (t, x) where t is the set s with x deleted. The set s must be non-empty.

See also Base.pop!, BangBang.pop!!.

SmallCollections.pop — Method

pop(s::S, x, default::T) where S <: SmallBitSet -> Tuple{S, Union{Int,T}}

If s contains x, return the pair (t, x) where t is the set s with x deleted. Otherwise return (s, default)

See also Base.pop!, BangBang.pop!!.

SmallCollections.delete — Function

delete(s::S, x) where S <: SmallBitSet -> S

If s contains x, return the set obtained by deleting that element. Otherwise return s.

See also Base.delete!, BangBang.delete!!.

BangBang support

If the package BangBang.jl is loaded, then the functions push, pop, delete, union, intersect, setdiff and symdiff for SmallBitSet as well as setindex, push, pushfirst, pop, popfirst and deleteat for SmallVector are also available in !!-form. For example, setindex!! with a SmallVector as first argument calls setindex. (For some reason, BangBang.jl does not implement insert!!.) Moreover, add!!(v::SmallVector, w::SmallVector) is a synonym for v+w.

This allows to write efficient code that works for both mutable and immutable arguments. For example, the function

f!!(v, ws...) = foldl(add!!, ws; init = v)

adds up its arguments, mutating the first argument v if possible.

Internal functions

SmallCollections.bitsize — Function

SmallCollections.bitsize(T::Type) -> Int

Return the size of the internal binary representation of T in bits.

See also Base.sizeof.

SmallCollections.default — Function

SmallCollections.default(::Type{T}) -> T

Return the default value of type T used for filling unused elements of a SmallVector. This must be defined as zero(T) if T supports algebraic operations. Otherwise it can be any value of type T.

This function has methods for number types, Char, String and Symbol. Methods for other types must be defined explicitly.

SmallCollections.top_set_bit — Function

SmallCollections.top_set_bit(x::Unsigned) -> Int

Return the position of the highest set bit in x (counting from 1), or return 0 if x is 0.

This function is analogous to Julia's internal function Base.top_set_bit, but it is also fast and correct for bit integers defined by BitIntegers.jl.

See also Base.top_set_bit.