Non-exported names

Public names

SmallCollections.bitsize — Function

SmallCollections.bitsize(T::Type) -> Int
SmallCollections.bitsize(x::T) where T -> Int

Return the size of the internal binary representation of T in bits. For Bool the function returns 1.

See also Base.sizeof.

SmallCollections.default — Function

SmallCollections.default(::Type{T}) where T -> T
SmallCollections.default(::T) where T -> T

Return the default value of type T used for filling unused elements of an AbstractSmallVector. This must be defined as zero(T) if T supports algebraic operations. Otherwise it can be any value of type T.

This function has methods for number types, bits types, Symbol, AbstractChar, AbstractString, AbstractFixedVector, AbstractSmallVector und SmallBitSet. Methods for other types must be defined explicitly.

See also Base.isbitstype.

SmallCollections.FixedVectorStyle — Type

SmallCollections.FixedVectorStyle <: Broadcast.AbstractArrayStyle{1}

The broadcasting style used for AbstractFixedVector.

See also AbstractFixedVector, Broadcast.AbstractArrayStyle.

SmallCollections.SmallVectorStyle — Type

SmallCollections.SmallVectorStyle <: Broadcast.AbstractArrayStyle{1}

The broadcasting style used for AbstractSmallVector.

See also AbstractSmallVector, Broadcast.AbstractArrayStyle.

Internal names

These names are not public and may change in future versions.

SmallCollections.element_type — Function

SmallCollections.element_type(itr) -> Type
SmallCollections.element_type(::Type) -> Type

Return the element type of an iterator or its type. This differs from eltype in that the element type of a Tuple or NamedTuple is determined via promote_type instead of promote_typejoin. For all other iterators there is no difference.

See also Base.eltype, Base.promote_type, Base.promote_typejoin.

Example

julia> eltype((1, 2, 3.0))
Real

julia> SmallCollections.element_type((1, 2, 3.0))
Float64

SmallCollections.AbstractBitInteger — Type

SmallCollections.AbstractBitInteger

This type is the union of Base.BitInteger, BitIntegers.AbstractBitSigned and BitIntegers.AbstractBitUnsigned.

SmallCollections.top_set_bit — Function

SmallCollections.top_set_bit(x::AbstractBitInteger) -> Int

Return the position of the highest set bit in x (counting from 1), or return 0 if x is 0.

This function is analogous to Julia's internal function Base.top_set_bit, but it is also fast and correct for bit integers defined by BitIntegers.jl.

See also Base.top_set_bit, SmallCollections.AbstractBitInteger.

SmallCollections.unsafe_shl — Function

SmallCollections.unsafe_shl(x::U, i::Integer) where U <: AbstractBitInteger -> U

This is a fast, but unsafe version of the left bit shift operator x << i. The shift i is assumed to be between 0 and bitsize(x)-1.

See also SmallCollections.bitsize, SmallCollections.AbstractBitInteger.

SmallCollections.unsafe_lshr — Function

SmallCollections.unsafe_lshr(x::U, i::Integer) where U <: AbstractBitInteger -> U

This is a fast, but unsafe version of the logical (or unsigned) right bit shift operator x >>> i. The shift i is assumed to be between 0 and bitsize(x)-1.

See also SmallCollections.bitsize, SmallCollections.AbstractBitInteger.

SmallCollections.pdep — Function

SmallCollections.pdep(x::Unsigned, y::U) where U <: Unsigned -> U

Assume that y has exactly m 1-bits. Then pdep(x, y) replaces these bits by the m lowest bits of x (in order) and returns the result. The remaining bits of x are ignored.

On x86_64 and i686 machines, this function uses the corresponding instruction from the BMI2 instruction set if possible. Without hardware support it is much slower.